Add navigation to user guide for help
The help feature shows the AB-3 user guide URL on the help window and
lets the user copy it with a button. The user can visit the webpage on
their browser using the URL on their clipboard.

As a step towards such simplification, let's navigate the user to the
Reache user guide on their default browser by clicking the 'Go to User
Guide' button on the existing help window.